tca642x: don't enable outputs until their value is set
The previous init procedure was a recipe for fried I/Os e.g. if they
were intended to be used in open drain mode tied together with other
open drain outputs.

It makes sure that if pins are requested to be inputs that they do
high-Z before the output data register is written, and more importantly
if pins are requested to be outputs it is done only after the output
data register has been written.

Previously all pins configured as "output low" would transition as
"high-Z -> output high -> output low" during init, which of course
is asking for trouble.

signed-off-by: Matthijs van Duin <matthijsvanduin@gmail.com>