GTA04: boot.scr: add "charge auto" command to enable USB power early.
This helps in the situation of a completely empty battery
where without this patch, charging would only begin after
Linux has set up the charger. This might be 30 or 50
seconds after first automatic start - with a very weakly
charged battery.

So we simply enable the charger with 500mA by the boot.scr
and then there is enough power from USB (it is <100mA if the
OMAP is off) to boot into Linux and continue charging.

This fixes issue: http://projects.goldelico.com/p/gta04-xloader/issues/279/

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>